Fixed crash when using Release version of PerformanceAPI
The return value of PerformanceAPI_EndEvent in the FFI interface was declared incorrectly (i.e. missing). This caused Rust and C to disagree about how much stack space should be reserved prior to calls to PerformanceAPI_EndEvent, which then caused a crash in PerformanceAPI_EndEvent when it tries to write the return value to the stack. Fixed by declaring the return value correctly.

Note that the return value is unused as it is only there to prevent calls to PerformanceAPI_EndEvent from optimized out due to tail call elimination; its contents are irrelevant. To prevent Rust users from having to deal with it, the `end_event` function internally ignores the return value.